+++ Sources/PhotoExhibition/Entities/Member.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Member: Hashable, Sendable, Identifiable, Codable {
   init(
     id: String, name: String? = nil, icon: String? = nil, icon_256x256: String? = nil,
-    icon_512x512: String? = nil, icon_1024x1024: String? = nil, createdAt: Date, updatedAt: Date
+    icon_512x512: String? = nil, icon_1024x1024: String? = nil, postAgreement: Bool = false, createdAt: Date, updatedAt: Date
   ) {
     self.id = id
     self.name = name
@@ -18,6 +18,7 @@ struct Member: Hashable, Sendable, Identifiable, Codable {
     self.icon_256x256 = icon_256x256
     self.icon_512x512 = icon_512x512
     self.icon_1024x1024 = icon_1024x1024
+    self.postAgreement = postAgreement
     self.createdAt = createdAt
     self.updatedAt = updatedAt
   }
@@ -28,6 +29,7 @@ struct Member: Hashable, Sendable, Identifiable, Codable {
   let icon_256x256: String?
   let icon_512x512: String?
   let icon_1024x1024: String?
+  let postAgreement: Bool
   let createdAt: Date
   let updatedAt: Date
 
@@ -45,6 +47,7 @@ struct Member: Hashable, Sendable, Identifiable, Codable {
     self.icon_256x256 = data["icon_256x256"] as? String
     self.icon_512x512 = data["icon_512x512"] as? String
     self.icon_1024x1024 = data["icon_1024x1024"] as? String
+    self.postAgreement = data["postAgreement"] as? Bool ?? false
     self.createdAt = createdAtTimestamp.dateValue()
     self.updatedAt = updatedAtTimestamp.dateValue()
   }
