+++ Sources/PhotoExhibition/Resources/Module.xcassets/gradient_blue.colorset/Contents.json
