+++ Sources/PhotoExhibition/Resources/Module.xcassets/gradient_purple.colorset/Contents.json
