+++ Sources/PhotoExhibition/Resources/Module.xcassets/illustration.imageset/Contents.json
