+++ Sources/PhotoExhibition/Resources/Module.xcassets/illustration.imageset/JOY.png
