+++ Sources/PhotoExhibition/Resources/Module.xcassets/illustration.imageset/JOY@2x.png
