+++ Sources/PhotoExhibition/Resources/Module.xcassets/illustration.imageset/JOY@3x.png
